# Conflicts:
#   frontend/src/components/user/UserCard.vue

+++ WebBlog_V1/src/main/java/com/web/blog/QnA/service/QnaService.java
@@ -2,6 +2,7 @@ package com.web.blog.QnA.service;
 
 import com.web.blog.Board.repository.PostRepository;
 import com.web.blog.Common.advice.exception.*;
+import com.web.blog.Common.model.Paging;
 import com.web.blog.Common.service.S3Service;
 import com.web.blog.Member.entity.Member;
 import com.web.blog.Member.repository.MemberRepository;
@@ -14,6 +15,7 @@ import com.web.blog.QnA.repository.ApostRepository;
 import com.web.blog.QnA.repository.QpostRepository;
 import com.web.blog.QnA.repository.QpostUploadsRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -71,15 +73,15 @@ public class QnaService {
     }
 
     //모든 질문글 검색
-    public List<Qpost> QuestionSearch(int which, long board_id, String keyword) { //검색: which = 1~4
+    public List<OnlyQpostMapping> QuestionSearch(int which, String keyword, Paging paging) { //검색: which = 1~4
         if (which == 1) { //제목 검색
-            return qpostRepository.findBySubjectContaining(keyword);
+            return qpostRepository.findBySubjectContaining(keyword, PageRequest.of(paging.getPageNo() - 1, Paging.COUNT_OF_PAGING_CONTENTS));
         } else if (which == 2) { //내용 검색
-            return qpostRepository.findByContentContaining(keyword);
+            return qpostRepository.findByContentContaining(keyword, PageRequest.of(paging.getPageNo() - 1, Paging.COUNT_OF_PAGING_CONTENTS));
         } else if (which == 3) { //작성자 검색
-            return qpostRepository.findByMember_NicknameContaining(keyword);
+            return qpostRepository.findByMember_NicknameContaining(keyword, PageRequest.of(paging.getPageNo() - 1, Paging.COUNT_OF_PAGING_CONTENTS));
         } else { //통합검색
-            return qpostRepository.findDistinctBySubjectContainingOrContentContainingOrMember_NicknameContaining(keyword, keyword, keyword);
+            return qpostRepository.findDistinctBySubjectContainingOrContentContainingOrMember_NicknameContaining(keyword, keyword, keyword, PageRequest.of(paging.getPageNo() - 1, Paging.COUNT_OF_PAGING_CONTENTS));
         }
     }
 
@@ -111,11 +113,12 @@ public class QnaService {
     //질문 삭제
     public boolean deleteQuestion(long qpost_id, Member member) {
         Qpost qpost = qpostRepository.findById(qpost_id).orElseThrow(CResourceNotExistException::new);
+        if(qpost.getAnswerCnt() > 0) throw new CAnsweredQuestionException();
         if(!member.getMsrl().equals(qpost.getMember().getMsrl())) {
             throw new CNotOwnerException();
         } else {
             if (qpost.getAnswerCnt() == 0) {
-                qTagService.deleteQtags(qpost);
+//                qTagService.deleteQtags(qpost);
                 qpostRepository.delete(qpost);
                 return true;
             } else if (qpost.getAnswerCnt() > 0) throw new CAnsweredQuestionException();
